Move handleFirebaseError to commons
